requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-filter==24.3
psycopg[binary]==3.2.10
python-dotenv==1.0.1
django-cors-headers==4.4.0
django-environ==0.11.2
whitenoise==6.7.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
django-debug-toolbar==4.4.6